refactor(flui-painting): split canvas/sugar.rs into 6 concern modules
Code-review finding #14. The Mythos chain U4 split moved the
sugar API out of the 3,305-LOC canvas.rs god module into a single
674-LOC `canvas/sugar.rs` file. That file mixed six unrelated
concerns (batch drawing, conditional draws, grid/repeat patterns,
debug visualization, convenience shapes, and the ~30-method chaining
API) behind one `pub mod sugar` declaration, which left the readers
of any one concern paying to scroll past five others.

This change replaces `canvas/sugar.rs` with `canvas/sugar/{mod,
batch, conditional, grid, debug, shapes, chain}.rs` -- one impl
Canvas block per concern, ~50-360 LOC each, with their own narrowly
scoped `use` lists. The public method surface is unchanged; every
method keeps its original signature and dispatches to the same
underlying primary methods.

Mapping:

- batch.rs       (5 methods, ~60 LOC):  draw_rects, draw_circles,
                                        draw_lines, draw_rrects,
                                        draw_paths
- conditional.rs (5 methods, ~65 LOC):  draw_rect_if, draw_circle_if,
                                        draw_if, draw_unless,
                                        draw_if_some
- grid.rs        (4 methods, ~70 LOC):  draw_grid, repeat_x,
                                        repeat_y, repeat_radial
- debug.rs       (4 methods, ~95 LOC):  debug_rect, debug_point,
                                        debug_axes, debug_grid
- shapes.rs      (4 methods, ~80 LOC):  draw_rounded_rect,
                                        draw_rounded_rect_corners,
                                        draw_pill, draw_ring
- chain.rs      (~30 methods, ~340 LOC): translated/scaled/rotated/
                                        rect/circle/text/.../also/
                                        when/when_else

Verified:
- cargo build -p flui-painting (default + --no-default-features)
- cargo test -p flui-painting (all binaries green, 0 regressions)
- cargo clippy -p flui-painting --lib --no-deps -- -D warnings
- bash scripts/port-check.sh (6/6 triggers clean)
- cargo check --workspace
